dating_site/requirements.txt
asgiref==3.7.2
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.6
flake8==6.0.0
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
pep8-naming==0.13.3
Pillow==10.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
